Add export json+tab data, add Quasar, rework the trench data fetch (#33)
**Issue** :  `https://github.com/archaiodata/thacer/issues/32`

**Description** :    
This PR is mainly for adding the export functionality in json and tab .
Because the export would be performed on all the trenches the fetch had
been improved. Rather than call each selected trenches on every change,
trenche is call one time and remove when needed.
The PR also introduce Quasar.
The PR remove the Burger.