Resolve Duplicate UUIDs with Temporary Trench Field and Map Display Improvements (#130)
**Issue** :  `https://github.com/esag-swiss/iDig-Webapp/issues/129`

**Description** :    
This PR addresses a critical concern related to UUIDs that may be
associated with multiple trenches. To resolve this issue, a temporary
"Trench" field is introduced during the pull operation, and it is
subsequently removed before saving (pushing) changes. This adjustment
ensures a unique identifier, combining both the "Trench" and the UUID.

Enhancements:

Improved search functionality: Introduces the ability to use operators
AND, OR, and select specific fields in the current language, enhancing
the search capabilities.
Trench selection by group: Adds the functionality to select trenches
based on group criteria, providing a more versatile and targeted
selection process.
Map Component Refactoring:

Additionally, this PR includes refactoring within the map component and
associated logic. The main focus of these changes is on optimizing the
display of overlays. This may involve enhancements, optimizations, or
structural improvements to the code, contributing to a better overall
functionality and user experience when visualizing overlays on the map.