Connections management (#171)
**Issue** :  `https://github.com/esag-swiss/iDig-Webapp/issues/170`

**Description** :    
This pull request includes significant updates to the `src/components`,
`src/services`, and `src/stores` directories to enhance the
functionality and user experience of the iDig webapp. The changes focus
on improving the header component, adding a profile component to manage
connections, and refining local storage management.

### Header and Profile Component Updates:
*
[`src/components/TheHeader.vue`](diffhunk://#diff-6c285e9a01c01158da00cf91e9cd2abaa011dbc91ee488de12b73d6d2742cac7L2-R25):
Replaced the traditional navbar with a `q-toolbar` and added a new
`TheHeaderProfile` component for managing user connections.
[[1]](diffhunk://#diff-6c285e9a01c01158da00cf91e9cd2abaa011dbc91ee488de12b73d6d2742cac7L2-R25)
[[2]](diffhunk://#diff-6c285e9a01c01158da00cf91e9cd2abaa011dbc91ee488de12b73d6d2742cac7L97-R69)
*
[`src/components/TheHeaderProfile.vue`](diffhunk://#diff-20bfdadd9465d8b9ba8ac636ebb4786a7186b4b8b7562e2de4bd179c4ccf64cbR1-R160):
Introduced a new component to handle user profiles, allowing users to
select or create connections and manage their credentials.

### Local Storage Enhancements:
*
[`src/services/localStorageManager.js`](diffhunk://#diff-43f8213e5a9663cf67eaea7d091b294de4801db53d013f504b41e756947a6dd9R10-R32):
Enhanced local storage functions to store and retrieve user connections
and project preferences, ensuring a smoother user experience.
[[1]](diffhunk://#diff-43f8213e5a9663cf67eaea7d091b294de4801db53d013f504b41e756947a6dd9R10-R32)
[[2]](diffhunk://#diff-43f8213e5a9663cf67eaea7d091b294de4801db53d013f504b41e756947a6dd9R98-R132)

### Codebase Simplification:
*
[`src/stores/data.js`](diffhunk://#diff-2db3f92e6ede07c64fb87d02cafbf7ef9fb8f3302f340d7e57838d8e0330eb3bL267-R271):
Removed the `fetchProjectTrenchesNamesFromFile` method and replaced
`fetchPreferences` with `fetchAndLoadPreferences` to streamline
preference management.
[[1]](diffhunk://#diff-2db3f92e6ede07c64fb87d02cafbf7ef9fb8f3302f340d7e57838d8e0330eb3bL267-R271)
[[2]](diffhunk://#diff-2db3f92e6ede07c64fb87d02cafbf7ef9fb8f3302f340d7e57838d8e0330eb3bL307-L338)

### Other Notable Changes:
*
[`src/router/index.js`](diffhunk://#diff-3210ca9ed615e6f0fc1743fc09be3c86bc38178dcc050e5a41ff7982d15875b7R2-R12):
Added a new route for the home path, linking it to the main `App`
component.