Add merged choice to sync conflicts (#175)
**Issue** :  `https://github.com/esag-swiss/iDig-Webapp/issues/174`

**Description** :    
This pull request includes several changes to the
`src/components/ThePatches.vue` file to improve the handling of patch
conflicts and user interactions. The most important changes include
updating the UI to provide better feedback to users, refactoring the
data handling logic for patch conflicts, and enhancing the
synchronization process with the server.

### UI Improvements:
* Added tooltips to buttons for better user guidance and changed button
classes for consistent styling.
* Updated conflict messages to be more descriptive and user-friendly.

### Data Handling Refactor:
* Replaced `patchesInConflict` with `UUIDsSyncPatches`, `itemsEdited`,
`totalPatches`, `keyModified`, and `finalVersionItems` to streamline the
conflict resolution logic.
* Introduced `compareObjectsThree` and `mergeVersions` methods to handle
three-way comparisons and merging of patch versions.

### Synchronization Enhancements:
* Modified the `syncSurvey` method to use `finalVersionItems` for
uploading, ensuring the correct version is synchronized.
* Improved server response handling to provide more informative
notifications to the user.
[[1]](diffhunk://#diff-4a9b2ddbf706cba248ada25e9e85f61be757f302529a82494a3a63896a576157L186-R221)
[[2]](diffhunk://#diff-4a9b2ddbf706cba248ada25e9e85f61be757f302529a82494a3a63896a576157L210-R246)

### Integration with Other Components:
* Added an event emitter to clear items in `TheTable.vue` when
synchronization is complete.