Fix Thetable sync bug (#198)
**Issue** :  `https://github.com/esag-swiss/iDig-Webapp/issues/195`

**Description** :    
This pull request refactors the codebase to improve modularity and
maintainability by centralizing survey-related logic, cleaning up unused
methods, and enhancing UI behavior. The changes primarily focus on
restructuring the `pushSurvey` functionality, removing redundant code,
and improving synchronization between components.

### Refactoring Survey Logic:

*
[`src/components/TheItem.vue`](diffhunk://#diff-c307e02209732f5a905ce468fbeb103f88bdda6b35c4275fd15736a893707f3bL565-L604):
Replaced the `pushSurvey` method with `pushSurveyHandler`, which
delegates the survey logic to the new `pushSurveyService`. This
centralizes the logic and simplifies the component.
*
[`src/services/pushSurveyService.js`](diffhunk://#diff-84647321b265aec34ef542fa11e913edbb783d6dd304501a7b828826360abcb1L50-R55):
Updated the `pushSurvey` function to handle synchronization actions
using the `useDataStore` instead of direct calls, improving modularity
and reducing coupling.

### Code Cleanup:

*
[`src/components/TheItem.vue`](diffhunk://#diff-c307e02209732f5a905ce468fbeb103f88bdda6b35c4275fd15736a893707f3bL501-L508):
Removed the unused `trenchtoUpdateWithoutTrenchProp` method and
commented-out code within `pushSurvey`, reducing clutter and improving
readability.

### UI Enhancements:

*
[`src/components/TheTable.vue`](diffhunk://#diff-350d67f884edbe3fcb2a11190a4940754ebcd9c5c7d519693978477a88272b3aL3-R3):
Modified the visibility logic for the item frame to display it when
either `selectedItem` or `syncPatches` is present, enhancing user
interaction.

### Minor Adjustments:

*
[`src/components/TheItem.vue`](diffhunk://#diff-c307e02209732f5a905ce468fbeb103f88bdda6b35c4275fd15736a893707f3bL22-R22):
Updated the `@click` handler for the upload button to use the new
`pushSurveyHandler` method.
*
[`src/services/pushSurveyService.js`](diffhunk://#diff-84647321b265aec34ef542fa11e913edbb783d6dd304501a7b828826360abcb1L17-R17):
Corrected the property name from `trench` to `Trench` in the
`removeTrenchProp` function to ensure consistency with the data
structure.